@@ -6,102 +6,94 @@ using System.Security.Cryptography;
 using LZ4;
 using NUnit.Framework;
 
-namespace LZ4Codec.Tests
+namespace LZ4Codec.Tests;
+
+public abstract class BaseTest
 {
-    public abstract class BaseTest
+    internal LZ4ServiceBase service;
+
+    /// <summary> low entropy data </summary>
+    protected Dictionary<string, byte[]> dataCompressable = new();
+
+    /// <summary> high entropy data </summary>
+    protected readonly Dictionary<string, byte[]> dataUncompressable = new();
+
+    [OneTimeSetUp]
+    public virtual void OneTimeSetup()
     {
-        internal LZ4ServiceBase service;
+        var buff = RandomNumberGenerator.GetBytes(64);
+        dataUncompressable.Add("random_very_small.bin", buff);
 
-        /// <summary> данные с низкой энтропией (сжимаемые) </summary>
-        protected Dictionary<string, byte[]> dataCompressable = new();
+        buff = RandomNumberGenerator.GetBytes(4096);
+        dataUncompressable.Add("random_small.bin", buff);
 
-        /// <summary> данные с высокой энтропией (несжимаемые) </summary>
-        protected readonly Dictionary<string, byte[]> dataUncompressable = new();
+        buff = RandomNumberGenerator.GetBytes(128 * 1024);
+        dataUncompressable.Add("random_big.bin", buff);
 
-        [OneTimeSetUp]
-        public virtual void OneTimeSetup()
+        var dataPath = Path.Combine(TestContext.CurrentContext.TestDirectory, "data");
+        dataCompressable = Directory.EnumerateFiles(dataPath, "*.*").ToDictionary(Path.GetFileName, File.ReadAllBytes);
+    }
+
+    [Test]
+    public void PackUnpackCompressable()
+    {
+        foreach (var data in dataCompressable)
         {
-            var rng = new RNGCryptoServiceProvider();
+            var copyOfOriginal = data.Value.ToArray();
 
-            var buff = new byte[64];
-            rng.GetBytes(buff);
-            dataUncompressable.Add("random_very_small.bin", buff);
+            var packed = service.Encode(copyOfOriginal);
+            Assert.True(data.Value.SequenceEqual(copyOfOriginal));
+            Assert.True(packed.Length > 0);
+            Assert.True(packed.Length <= copyOfOriginal.Length);
 
-            buff = new byte[4096];
-            rng.GetBytes(buff);
-            dataUncompressable.Add("random_small.bin", buff);
-
-            buff = new byte[128 * 1024];
-            rng.GetBytes(buff);
-            dataUncompressable.Add("random_big.bin", buff);
-
-            var dataPath = Path.Combine(TestContext.CurrentContext.TestDirectory, "data");
-            dataCompressable = Directory.EnumerateFiles(dataPath, "*.*").ToDictionary(Path.GetFileName, File.ReadAllBytes);
+            var unpacked = service.Decode(packed);
+            Assert.IsTrue(unpacked.Length > 0);
+            Assert.True(unpacked.SequenceEqual(copyOfOriginal));
         }
-        
-        [Test]
-        public void PackUnpackCompressable()
+    }
+
+    [Test]
+    public void PackUnpackUncompressable()
+    {
+        foreach (var data in dataUncompressable)
         {
-            foreach (var data in dataCompressable)
-            {
-                var copyOfOriginal = data.Value.ToArray();
+            var copyOfOriginal = data.Value.ToArray();
 
-                var packed = service.Encode(copyOfOriginal);
-                Assert.True(data.Value.SequenceEqual(copyOfOriginal));
-                Assert.True(packed.Length > 0);
-                Assert.True(packed.Length <= copyOfOriginal.Length);
+            var packed = service.Encode(copyOfOriginal);
+            Assert.True(data.Value.SequenceEqual(copyOfOriginal));
+            Assert.True(packed.Length     > 0);
+            Assert.True(packed.Length - 8 <= copyOfOriginal.Length);
 
-                var unpacked = service.Decode(packed);
-                Assert.IsTrue(unpacked.Length > 0);
-                Assert.True(unpacked.SequenceEqual(copyOfOriginal));
-            }
+            var unpacked = service.Decode(packed);
+            Assert.IsTrue(unpacked.Length > 0);
+            Assert.True(unpacked.SequenceEqual(copyOfOriginal));
         }
+    }
 
-        [Test]
-        public void PackUnpackUncompressable()
-        {
-            foreach (var data in dataUncompressable)
-            {
-                var copyOfOriginal = data.Value.ToArray();
+    [Test]
+    public void EmptyPack()
+    {
+        var packed = service.Encode(Array.Empty<byte>());
+        Assert.IsTrue(packed.Length == 0);
+    }
 
-                var packed = service.Encode(copyOfOriginal);
-                Assert.True(data.Value.SequenceEqual(copyOfOriginal));
-                Assert.True(packed.Length > 0);
-                Assert.True(packed.Length - 8 <= copyOfOriginal.Length);
+    [Test]
+    public void EmptyUnpack()
+    {
+        var unpacked = service.Decode(Array.Empty<byte>());
+        Assert.IsTrue(unpacked.Length == 0);
+    }
 
-                var unpacked = service.Decode(packed);
-                Assert.IsTrue(unpacked.Length > 0);
-                Assert.True(unpacked.SequenceEqual(copyOfOriginal));
-            }
-        }
+    [Test]
+    public void InvalidUnpack()
+    {
+        var buff = RandomNumberGenerator.GetBytes(1024);
+        BitConverter.TryWriteBytes(buff.AsSpan(),          -2);
+        BitConverter.TryWriteBytes(buff.AsSpan().Slice(4), -3);
+        Assert.Throws<InvalidOperationException>(() => service.Decode(buff));
 
-        [Test]
-        public void EmptyPack()
-        {
-            var packed = service.Encode(Array.Empty<byte>());
-            Assert.IsTrue(packed.Length == 0);
-        }
-
-        [Test]
-        public void EmptyUnpack()
-        {
-            var unpacked = service.Decode(Array.Empty<byte>());
-            Assert.IsTrue(unpacked.Length == 0);
-        }
-
-        [Test]
-        public void InvalidUnpack()
-        {
-            var buff = new byte[1024];
-            var rng  = new RNGCryptoServiceProvider();
-            rng.GetBytes(buff);
-            BitConverter.TryWriteBytes(buff.AsSpan(), -2);
-            BitConverter.TryWriteBytes(buff.AsSpan().Slice(4), -3);
-            Assert.Throws<InvalidOperationException>(() => service.Decode(buff));
-            
-            BitConverter.TryWriteBytes(buff.AsSpan(), 123);
-            BitConverter.TryWriteBytes(buff.AsSpan().Slice(4), 444);
-            Assert.Throws<InvalidOperationException>(() => service.Decode(buff));
-        }
+        BitConverter.TryWriteBytes(buff.AsSpan(),          123);
+        BitConverter.TryWriteBytes(buff.AsSpan().Slice(4), 444);
+        Assert.Throws<InvalidOperationException>(() => service.Decode(buff));
     }
 }
