@@ -1,16 +1,14 @@
-using System.Linq;
 using LZ4;
 using NUnit.Framework;
 
-namespace LZ4Codec.Tests
+namespace LZ4Codec.Tests;
+
+public class x32 : BaseTest
 {
-    public class x32 : BaseTest
+    [OneTimeSetUp]
+    public override void OneTimeSetup()
     {
-        [OneTimeSetUp]
-        public override void OneTimeSetup()
-        {
-            base.OneTimeSetup();
-            service = new LZ4Service32();
-        }
+        base.OneTimeSetup();
+        service = new LZ4Service32();
     }
 }
