@@ -1,15 +1,14 @@
 using LZ4;
 using NUnit.Framework;
 
-namespace LZ4Codec.Tests
+namespace LZ4Codec.Tests;
+
+public class x64 : BaseTest
 {
-    public class x64 : BaseTest
+    [OneTimeSetUp]
+    public override void OneTimeSetup()
     {
-        [OneTimeSetUp]
-        public override void OneTimeSetup()
-        {
-            base.OneTimeSetup();
-            service = new LZ4Service64();
-        }
+        base.OneTimeSetup();
+        service = new LZ4Service64();
     }
 }
