@@ -1,83 +1,90 @@
 ﻿using System;
-using System.Diagnostics;
+using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Security.Cryptography;
+using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Jobs;
+using BenchmarkDotNet.Order;
+using BenchmarkDotNet.Reports;
+using BenchmarkDotNet.Running;
+using LZ4;
 
-namespace LZ4
+var summary = BenchmarkRunner.Run<MainTest>();
+
+[SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+[WarmupCount(3)]
+[IterationCount(3)]
+[MemoryDiagnoser]
+[Config(typeof(Config))]
+public class MainTest
 {
-    class Program
+    readonly string            Location = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
+    Dictionary<string, byte[]> rawFiles;
+    Dictionary<string, byte[]> packedFiles;
+
+    [Params("___1KB", "__16KB", "_128KB", "_512KB", "1024KB")]
+    public string FileName { get; set; }
+
+    [Params(32, 64)]
+    public int Bitness { get; set; }
+
+    [GlobalSetup]
+    public void GlobalSetup()
     {
-        static readonly string Location = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
+        rawFiles = Directory.EnumerateFiles(Path.Combine(Location, "data"), "*.*")
+                            .ToDictionary(Path.GetFileNameWithoutExtension, File.ReadAllBytes);
+        packedFiles = rawFiles.ToDictionary(p => p.Key, p => p.Value.AsSpan().PackLZ4().ToArray());
+    }
 
-        static void Main(string[] args)
+    [Benchmark]
+    public int Pack()
+    {
+        if (Bitness == 32) LZ4Codec.Use32Version();
+        else LZ4Codec.Use64Version();
+        return rawFiles[FileName].AsSpan().PackLZ4().Length;
+    }
+
+    [Benchmark]
+    public int Unpack()
+    {
+        if (Bitness == 32) LZ4Codec.Use32Version();
+        else LZ4Codec.Use64Version();
+        return packedFiles[FileName].AsSpan().UnpackLZ4().Length;
+    }
+
+    #region Config
+
+    class Config : ManualConfig
+    {
+        public Config() => Orderer = new CustomOrderer();
+
+        private class CustomOrderer : IOrderer
         {
-            var rng = new RNGCryptoServiceProvider();
-            var files = Directory.EnumerateFiles(Path.Combine(Location, "test"), "*.*")
-                                 .ToDictionary(Path.GetFileName, File.ReadAllBytes);
+            public IEnumerable<BenchmarkCase> GetExecutionOrder(ImmutableArray<BenchmarkCase> benchmarksCase, IEnumerable<BenchmarkLogicalGroupRule> order = null) =>
+                benchmarksCase.OrderBy(p => p.Descriptor.WorkloadMethod.Name)
+                              .ThenBy(p => (string) p.Parameters["FileName"])
+                              .ThenBy(p => (int) p.Parameters["Bitness"]);
 
-            {
-                // pack speed
-                var sw           = Stopwatch.StartNew();
-                int unpackLength = 0;
-                int packLength   = 0;
-                for (var i = 0; i < 100; i++)
-                {
-                    foreach (var item in files)
-                    {
-                        var packed = item.Value.AsSpan().PackLZ4();
-                        unpackLength += item.Value.Length;
-                        packLength   += packed.Length;
-                        //var unpacked = packed.UnpackLZ4();
-                    }
-                }
-            
-                // x32 - 583 MB/sec
-                // x64 - 546 MB/sec
-                Console.WriteLine("{0:F2}", unpackLength / 1_048_576.0 / sw.Elapsed.TotalSeconds);
-            }
-            
-            {
-                // unpack speed
-                var sw           = Stopwatch.StartNew();
-                int unpackLength = 0;
-                int packLength   = 0;
-                var packedFiles = files.ToDictionary(p => p.Key, p => p.Value.AsSpan().PackLZ4().ToArray());
-                
-                for (var i = 0; i < 100; i++)
-                {
-                    foreach (var item in packedFiles)
-                    {
-                        var unpacked = item.Value.AsSpan().UnpackLZ4();
-                        unpackLength += unpacked.Length;
-                        packLength   += item.Value.Length;
-                    }
-                }
-            
-                // x32 - 762 MB/sec
-                // x64 - 800 MB/sec
-                Console.WriteLine("{0:F2}", unpackLength / 1_048_576.0 / sw.Elapsed.TotalSeconds);
-            }
+            public IEnumerable<BenchmarkCase> GetSummaryOrder(ImmutableArray<BenchmarkCase> benchmarksCase, Summary summary) =>
+                benchmarksCase.OrderBy(p => p.Descriptor.WorkloadMethod.Name)
+                              .ThenBy(p => (string) p.Parameters["FileName"])
+                              .ThenBy(p => (int) p.Parameters["Bitness"]);
 
-            // var buff = new byte[3072];
-            // rng.GetBytes(buff);
-            // files.Add("random_small.bin", buff);
-            //
-            // buff = new byte[100_000];
-            // rng.GetBytes(buff);
-            // files.Add("random_big.bin", buff);
-            //
-            // foreach (var item in files)
-            // {
-            //     Console.WriteLine("{0}", item.Key);
-            //
-            //     var packed   = item.Value.PackLZ4();
-            //     var unpacked = packed.UnpackLZ4();
-            //
-            //     if (!item.Value.SequenceEqual(unpacked))
-            //         throw new InvalidDataException(item.Key);
-            // }
+            public string GetHighlightGroupKey(BenchmarkCase benchmarkCase) => null;
+
+            public string GetLogicalGroupKey(ImmutableArray<BenchmarkCase> allBenchmarksCases, BenchmarkCase benchmarkCase) =>
+                benchmarkCase.Job.DisplayInfo + "_" + benchmarkCase.Parameters.DisplayInfo;
+
+            public IEnumerable<IGrouping<string, BenchmarkCase>> GetLogicalGroupOrder(IEnumerable<IGrouping<string, BenchmarkCase>> logicalGroups,
+                                                                                      IEnumerable<BenchmarkLogicalGroupRule>        order = null) =>
+                logicalGroups.OrderBy(it => it.Key);
+
+            public bool SeparateLogicalGroups => false;
         }
     }
+
+    #endregion
 }
